+++ Executor/Executor-Threshold.hpp
@@ -16,7 +16,7 @@ class Threshold {
 public:
   Threshold(
     uint sampleRate,
-    std::function<float()> func
+    std::function<double()> func
   )
   : buffer_(Data::Buffer(sampleRate / 100)) // 10ms length
   , writeLoc_(0uz)
@@ -24,9 +24,9 @@ public:
   , func_(func)
   {}
 
-  float handle() {
+  double handle() {
     if (isActive_) {
-      float outputSample = func_();
+      double outputSample = func_();
       buffer_[writeLoc_] = outputSample;
       writeLoc_ = ++writeLoc_ % buffer_.size();
       if (buffer_.rms() < 0.001f) {
@@ -46,7 +46,7 @@ private:
   Data::Buffer buffer_;
   size_t writeLoc_;
   bool isActive_;
-  std::function<float()> func_;
+  std::function<double()> func_;
 }; // Threshold class
 
 } // MilletDSP::Executor namespace
